restructure files/folders for print components
