+++ frontend/src/components/print/config/Activity.vue
