+++ frontend/src/components/print/config/Picasso.vue
