+++ frontend/src/components/print/print-react/components/picasso/DayHeader.jsx
@@ -1,7 +1,7 @@
 // eslint-disable-next-line no-unused-vars
 import React from 'react'
 import { View, Text } from '../../reactPdf.js'
-import dayjs from '../../../../../../common/helpers/dayjs.js'
+import dayjs from '@/common/helpers/dayjs.js'
 
 const wrapperStyles = {
   flexGrow: '1',
