+++ frontend/src/components/print/print-react/components/picasso/Responsibles.jsx
