+++ frontend/src/components/print/print-react/components/picasso/ScheduleEntry.jsx
