+++ frontend/src/components/print/print-react/components/scheduleEntry/ScheduleEntry.jsx
@@ -3,7 +3,7 @@ import React from 'react'
 import { View, Text } from '../../reactPdf.js'
 import styles from '../styles.js'
 import Responsibles from '../picasso/Responsibles.jsx'
-import dayjs from '../../../../../../common/helpers/dayjs.js'
+import dayjs from '@/common/helpers/dayjs.js'
 import ContentNode from './contentNode/ContentNode.jsx'
 
 const fontSize = 8
