+++ frontend/src/components/print/print-react/components/scheduleEntry/contentNode/LAThematicArea.jsx
