+++ frontend/src/components/print/print-react/components/scheduleEntry/contentNode/Notes.jsx
