+++ frontend/src/components/print/print-react/minimalHalJsonVuex.js
